fix: prepare SPEED-Bench in the runner before driver.run
